requirements.txt
absl-py==2.3.1
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
clarabel==0.9.0
cloudpickle==3.1.1
cmake==4.1.2
contourpy==1.1.1
cryptography==39.0.1
cvxpy==1.4.1
cycler==0.12.1
django==4.2.24
easydict==1.10
ecos==2.0.14
einops==0.8.1
ffmpeg-python==0.2.0
filelock==3.12.2
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.4
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
lit==18.1.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
ofa==0.1.0.post202307202001
opencv-python==4.7.0.72
osqp==1.0.5
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.5
psutil==7.0.0
ptflops==0.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybind11==3.0.1
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
scs==3.2.7.post2
six==1.16.0
sqlparse==0.5.3
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tifffile==2023.7.10
timm==0.9.2
torchvision==0.15.2
tqdm==4.67.1
triton==2.0.0
typing-extensions==4.3.0
tzdata==2025.2
urllib3==1.26.14
werkzeug==3.0.6
zipp==3.20.2
